refactor(video_editor): move ClipWaveform into the models layer
The overlay widget only needs the waveform data, but importing it from
the manager pulled a service import into the UI layer, which the
ui_service_boundary guard rejects for new files. The type is data, not a
service, so it belongs in models — the widget imports the model and the
manager keeps the extraction.
